third_party/blink/web_tests/fast/js/mozilla/strict/15.4.4.9-expected.txt

PASS 'use strict'; var a = arr(); [a.shift(), a] threw exception of type TypeError.
PASS var a = arr(); [a.shift(), a] threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var o = obj(); [Array.prototype.shift.call(o), o] threw exception of type TypeError.
PASS var o = obj(); [Array.prototype.shift.call(o), o] threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var a = agap(); [a.shift(), a] threw exception of type TypeError.
PASS var a = agap(); [a.shift(), a] threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var o = ogap(); [Array.prototype.shift.call(o), o] threw exception of type TypeError.
PASS var o = ogap(); [Array.prototype.shift.call(o), o] threw exception of type TypeError.
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
